Add path methods from UriInterface
Add methods getPath and WithPath from PSR-7 UriInterface.

Therefore also add a new method resolvePath to the Resolver, that only resolves a path against a base path instead of anything against a base url.

Also improve path validation to just percent-encode (or remove) any character that isn't allowed within a path according to RFC 3986 and adapt some tests that change because of that.