Add classes Host and Domain
An instance of Host is used in the Url class instead of the host string property. An instance of Host creates an instance of Domain when the host string ends with a public suffix.
Validation of host components still happens in the Url class, as it has a Validator instance. The Host and Domain classes are rather simple and assume that Validation happens somewhere else.
Nonetheless this change simplifies and centralizes handling of the host and the domain parts and reduces the size of the Url and Helper classes.