Fix things according to UriInterface docs
As the phpdocs in the UriInterface state that the get methods (except port) must return an empty string, the test cases are changed to explicitly check for an empty string, not only for empty.

Fix double encoding of already percent encoded characters in path, query and fragment.

Also fix building the full Url in the special case when no authority is present and the current path starts with more than one slash. In this case the starting slashes in the path are reduced to one as it says in the UriInterface documentation.