Refactored and improved validation
Split url and absoluteUrl validation methods to one that returns the valid url as a string and another one that returns the valid url and all valid components as an array.

Add validation for authority and userInfo both with the same two methods to either get a valid string or an array of valid components.

Also separate validation of domainLabel from the domain method.

Add a method callValidationByComponentName to avoid calling validation methods in a loop with self::{$componentName} which makes calls to a method untraceable for IDEs.