Fix PHP 7.2/7.3 issue with idn_to methods (#12)
Add wrapper methods for idn_to_ascii and idn_to_utf8 in Helpers, because
the argument variant in these methods has the default value
INTL_IDNA_VARIANT_2003 until PHP 7.4 but that value was deprecated in
PHP 7.2. PHP 7.4 defaults to INTL_IDNA_VARIANT_UTS46 which is used in
the new helper methods.