Support PHP 8.0 and improve getting default ports (#19)
Change in Validator for PHP 8.0, because output of PHP's parse_url is
different than in PHP 7 when an url includes a delimiter for query or
fragment but has no actual query or fragment (followed by empty string).

Change PHP version requirement in composer.json.

Temporarily add PHP_CS_FIXER_IGNORE_ENV=1 to `composer cs` command until
PHP Coding Standards Fixer fully supports PHP 8.0.

PHP's getservbyname() function uses the /etc/services file and when it's
missing the function returns false for all schemes. Fix that by having
a list within the package.
Also refactor the abstract List Updater because it was built to fetch
lists from files from the web. The DefaultPorts list only uses the
schemes list from the package and tries to get the default ports either
from the current list in the package or PHP's getservbyname() function
(via the Helpers::getStandardPortByScheme() method).
Further add a get method to the abstract List Store class so the
DefaultPorts class can get the default port for a scheme and not only
check if it exists.