Add queryString method using crwlr/query-string
When installed the new queryString method provides advanced access to
the url query via the Query class from the new crwlr/query-string
package. The new package requires PHP version 8, and the url package
requires only 7.2. If there will be another major version of the url
package it will at least require 8.0 and require the query-string
package. For now it's only suggested. The method throws an exception
when the query-string package isn't installed.